constructor for range partitioning schema

+++ src/lib/storage/partitioning/round_robin_partition_schema.hpp
@@ -8,11 +8,11 @@ namespace opossum {
 class RoundRobinPartitionSchema : public PartitionSchema {
 
 public:
-	RoundRobinPartitionSchema(int number_of_partitions);
+	RoundRobinPartitionSchema(size_t number_of_partitions);
 
 protected:
-    int _number_of_partitions;
-    PartitionID _next_partition;
+  int _number_of_partitions;
+  PartitionID _next_partition;
 
 };
